Creation
GUIProject.java is the basic interface for adding projects as well as sorting them in the five various ways.
Project.java is the constructor for the projects, and builds them with their information.
Status.java is the constructor for the status of the project, since there are only 3 statuses. I also gave each status a specific number so that they can be easily sorted.

The sorting algorithm is selection sort and I have used ArrayLists to contain the list of projects, it may be easier to use ArrayLists for the final version of the project.

As of right now, each of the sorting algorithms sorts them in numeric order/alphabetical order. When we implement this into the GUI code, I can make it so it does both, and we can have two buttons, one to sort in ascending order and another to sort in descending order.

Status right now is it's own class, but we can make status a drop-down list later that will basically do the same thing as the class I have now, I have just included it for sake of this interface. As a result, when entering status, it will only take either "Not Started", "Completed", or "In Progress". Sorting due dates right now is just a basic number comparison that I can change later once the drop-down for dates are selected.  

Feel free to ask me if there's any questions for clarification.